+++ interface/src/js/indexFunctions.js
@@ -659,42 +659,3 @@ export function selectChargeRaids(movelist, moveTable, pokName, pokTable) {
 
     return primaryName
 }
-
-export function returnRateStyle(rate) {
-    if (rate >= 630) {
-        return ["+2", "4"]
-    }
-    if (rate > 500) {
-        return ["+1", "3"]
-    }
-    if (rate === 500) {
-        return ["0", "0"]
-    }
-    if (rate >= 370) {
-        return ["-1", "1"]
-    }
-    return ["-2", "2"]
-}
-
-export function returnVunStyle(rate) {
-    if (rate > 1.600) {
-        return "2"
-    }
-    if (rate > 1.000) {
-        return "1"
-    }
-    if (rate === "1.000") {
-        return "0"
-    }
-    if (rate > 0.391) {
-        return "3"
-    }
-    return "4"
-}
-
-
-
-
-
-
-
